requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.22.0
mlflow==2.20.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
matplotlib==3.10.0
seaborn==0.13.2
jupyterlab==4.3.4
scikit-learn==1.6.1
ipython
ipykernel
